moved annotations, data_file to /data
