+++ medacy/tools/__init__.py
@@ -1,3 +1,1 @@
 from medacy.tools.converters import brat_to_con, con_to_brat
-from medacy.tools.annotations import Annotations
-from medacy.tools.data_file import DataFile
